♻️  refactor: Create decorators, lib and people folders to contain files of their type
Split base_decorator, add capitalize_decorator and trimmer_decorator
