+++ app/app.rb
@@ -1,8 +1,8 @@
-require_relative './book'
-require_relative './student'
-require_relative './teacher'
-require_relative './rental'
-require_relative './person'
+require_relative './lib/book'
+require_relative './lib/rental'
+require_relative './people/person'
+require_relative './people/student'
+require_relative './people/teacher'
 require_relative './helpers/listing_methods'
 require_relative './helpers/validations'
 require_relative './helpers/create_methods'
