+++ app/lib/book.rb
