+++ app/lib/classroom.rb
