+++ app/lib/rental.rb
