+++ app/people/nameable.rb
