+++ app/people/person.rb
