+++ app/people/student.rb
