+++ app/people/teacher.rb
